# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Dealer Operator API
  version: 4.0.0
  description: |
    Contract for the operator publish console and dealer inventory APIs.
    Handlers are implemented in Fastify; this spec is the source for the generated web client.
servers:
  - url: /
    description: Same-origin (Vite proxy in dev)
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      x-route-classification: auth
      operationId: operatorLogin
      summary: Operator login — issues op_session HttpOnly cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorIdentityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/me:
    get:
      tags: [Auth]
      security:
        - CookieAuth: []
      x-route-classification: auth
      operationId: getOperatorMe
      summary: Get current operator identity
      responses:
        '200':
          description: Authenticated operator identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorIdentityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/dealers/{dealershipId}/notifications:
    get:
      tags: [Leads]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: listDealerNotifications
      summary: List dealer notifications
      description: |
        Returns DealerNotification rows for this dealership, most recent first.
        Covers LEAD_CAPTURED and VEHICLE_SOLD_MARKETPLACE events. deliveryStatus
        reflects email delivery (PENDING | SENT | FAILED); this endpoint is the
        operator read surface regardless of email outcome.
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: type
          in: query
          description: Filter by notification type (e.g. VEHICLE_SOLD_MARKETPLACE)
          schema: { type: string }
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerNotificationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/dealers/{dealershipId}/notification-channels:
    get:
      tags: [Leads]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getNotificationChannels
      summary: Get notification channel configuration
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Channel configuration
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Leads]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: updateNotificationChannels
      summary: Update notification channel configuration
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated channel configuration
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/dealers/{dealershipId}/buyer-outreach:
    get:
      tags: [Leads]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: listBuyerOutreach
      summary: List auto-response messages sent to buyers
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: channel
          in: query
          schema: { type: string, enum: [email, sms] }
      responses:
        '200':
          description: Outreach history
          content:
            application/json:
              schema:
                type: object
                required: [outreach, total]
                properties:
                  outreach:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/buyer-outreach-stats:
    get:
      tags: [Admin]
      security:
        - OperatorAuth: []
      x-route-classification: admin
      operationId: getAdminBuyerOutreachStats
      summary: System-wide buyer auto-response usage stats (SUPER_ADMIN only)
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 90, default: 30 }
      responses:
        '200':
          description: Outreach stats
          content:
            application/json:
              schema:
                type: object
                required: [days, totalSent, totalFailed, byChannel, topDealers]
                properties:
                  days:        { type: integer }
                  totalSent:   { type: integer }
                  totalFailed: { type: integer }
                  byChannel:
                    type: array
                    items:
                      type: object
                      required: [channel, status, count]
                      properties:
                        channel: { type: string }
                        status:  { type: string }
                        count:   { type: integer }
                  topDealers:
                    type: array
                    items:
                      type: object
                      required: [dealershipId, sent]
                      properties:
                        dealershipId: { type: string }
                        sent:         { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    OperatorAuth:
      type: apiKey
      in: cookie
      name: op_session
      description: |
        Operator session cookie issued by POST /api/auth/login (8-hour lifetime, HttpOnly, Secure in production).
        In development/test the Fastify guard also accepts an x-operator-id request header as a fallback;
        that path is never active when NODE_ENV=production.
    CookieAuth:
      type: apiKey
      in: cookie
      name: op_session
      description: HttpOnly session cookie issued by POST /api/auth/login. Secure flag set in production.
  parameters:
    dealershipId:
      name: dealershipId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Operator authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Operator does not have access to this dealership
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    OperatorLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
    OperatorIdentityResponse:
      type: object
      required: [id, email, role, dealerAccessIds]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, OPERATOR, DEALER_OPERATOR]
        dealerAccessIds:
          type: array
          items:
            type: string
          description: Dealership IDs the operator can access. Empty for SUPER_ADMIN (global access).
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DealerNotificationListResponse:
      type: object
      required: [notifications, total]
      properties:
        total:
          type: integer
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/DealerNotificationItem'
    DealerNotificationItem:
      type: object
      required: [id, type, payload, deliveryStatus, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          description: Notification kind, e.g. LEAD_CAPTURED or VEHICLE_SOLD_MARKETPLACE
        payload:
          type: object
          additionalProperties: true
        deliveryStatus:
          type: string
          enum: [PENDING, SENT, FAILED]
          description: Email delivery outcome — not an operator read/unread state
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...